requirements.txt
google-api-python-client==2.108.0
notion-client==2.1.0
python-dotenv==1.0.0
textblob==0.17.1
janome==0.5.0
